Fixed Client Within the Unified Project >> Added Backup Zips for SQL Files
Added DLLs for SQL inside of the project folder so it will compile and
run.

+++ New Implementation/ManaCraft/LoginPopup.h
@@ -36,7 +36,7 @@ public:
 		ClientAPI::GetGuiContainer("LoginPopup")->SubscribeOnEnterKeyPressed(LoginPopup::ButtonPress_Enter);
 
 		//-- Set Menu Position
-		ClientAPI::GetGuiContainer("LoginPopup")->SetPosition(APIHelper::RectHelper(SCREEN_WIDTH / 2.9, SCREEN_HEIGHT / 2.5, 0, 0));
+		ClientAPI::GetGuiContainer("LoginPopup")->SetPosition(APIHelper::RectHelper((int)(SCREEN_WIDTH / 2.9), (int)(SCREEN_HEIGHT / 2.5), 0, 0));
 	
 		// Turn off
 		ClientAPI::GetGuiContainer("LoginPopup")->Active = false;
